+++ GUI/src/pages/Training/Intents/index.tsx
@@ -1,7 +1,6 @@
-import { FC, useCallback, useEffect, useState } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { FC, useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation } from '@tanstack/react-query';
 import * as Tabs from '@radix-ui/react-tabs';
 import { AxiosError } from 'axios';
 
@@ -12,68 +11,25 @@ import LoadingDialog from '../../../components/LoadingDialog';
 import withAuthorization, { ROLES } from 'hoc/with-authorization';
 import IntentTabList from './IntentTabList';
 import IntentDetails from './IntentDetails';
-import {
-  intentResponseToIntent,
-  IntentsWithExamplesCountResponse,
-  IntentWithExamplesCount,
-  IntentWithExamplesCountResponse,
-} from 'types/intentWithExampleCounts';
+import { useIntentsData } from './useIntentsData';
 
 const Intents: FC = () => {
   const { t } = useTranslation();
-  const queryClient = useQueryClient();
   const toast = useToast();
-  const [searchParams] = useSearchParams();
 
-  const [intents, setIntents] = useState<IntentWithExamplesCount[]>([]);
-  const [selectedIntent, setSelectedIntent] = useState<IntentWithExamplesCount | null>(null);
-  const [refreshing, setRefreshing] = useState(false);
-  const [filter, setFilter] = useState('');
-
-  const { data: intentsResponse, isLoading } = useQuery<IntentsWithExamplesCountResponse>({
-    queryKey: ['intents/with-examples-count'],
+  const { intents, selectedIntent, setSelectedIntent, queryRefresh, isLoading } = useIntentsData({
+    queryKey: 'intents/with-examples-count',
   });
 
-  useEffect(() => {
-    if (intentsResponse) {
-      setIntents(intentsResponse.response.intents.map((intent) => intentResponseToIntent(intent)));
-    }
-  }, [intentsResponse]);
-
-  const queryRefresh = useCallback(
-    async (newIntent?: string) => {
-      const response = await queryClient.fetchQuery<IntentsWithExamplesCountResponse>(['intents/with-examples-count']);
-
-      if (response) {
-        setIntents(response.response.intents.map((intent) => intentResponseToIntent(intent)));
-
-        const selectedIntent = response.response.intents.find(
-          (intent) => intent.id === newIntent
-        ) as IntentWithExamplesCountResponse;
-
-        setSelectedIntent(intentResponseToIntent(selectedIntent));
-      }
-    },
-    [queryClient]
-  );
-
-  useEffect(() => {
-    let intentParam = searchParams.get('intent');
-    if (!intentParam) return;
-
-    const queryIntent = intents.find((intent) => intent.id === intentParam);
-
-    if (queryIntent) {
-      setSelectedIntent(queryIntent);
-    }
-  }, [intents, searchParams]);
+  const [refreshing, setRefreshing] = useState(false);
+  const [filter, setFilter] = useState('');
 
   const handleTabsValueChange = useCallback(
     (value: string) => {
       const selectedIntent = intents.find((intent) => intent.id === value);
       setSelectedIntent(selectedIntent!);
     },
-    [intents]
+    [intents, setSelectedIntent]
   );
 
   const newIntentMutation = useMutation({
